rename: global 디렉토리 내 클래스 디렉토리 분리
