+++ packy-domain/src/main/java/com/dilly/auth/domain/AppleAccount.java
@@ -1,6 +1,6 @@
 package com.dilly.auth.domain;
 
-import com.dilly.global.BaseTimeEntity;
+import com.dilly.global.domain.BaseTimeEntity;
 import com.dilly.member.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
