+++ packy-domain/src/main/java/com/dilly/auth/domain/KakaoAccount.java
@@ -1,6 +1,6 @@
 package com.dilly.auth.domain;
 
-import com.dilly.global.BaseTimeEntity;
+import com.dilly.global.domain.BaseTimeEntity;
 import com.dilly.member.domain.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
