+++ packy-domain/src/main/java/com/dilly/global/constant/Constants.java
@@ -1,4 +1,4 @@
-package com.dilly.global;
+package com.dilly.global.constant;
 
 public class Constants {
     private Constants() {}
