+++ packy-domain/src/main/java/com/dilly/global/formatter/P6SpyFormatter.java
@@ -1,4 +1,4 @@
-package com.dilly.global;
+package com.dilly.global.formatter;
 
 import com.p6spy.engine.logging.Category;
 import com.p6spy.engine.spy.P6SpyOptions;
